Move esql translators into their own class
